differentiate between rate conversion factors
There are actually three rates encodings of interest. One is the
atomic ratio, which is quoteQty/baseQty. Quantities for this rate
are in atomic units. Then there is the message-rate encoding, which
is roughly uint64(atomicRatio * 1e8). Then there is the
conventional ratio, which is similar to the atomic ratio, but in
conventional units. The relationship between the atomic ratio and
the conventional ratio is

conventionalRatio = atomicRatio * (baseConversionFactor / quoteConversionFactor)

where the conversion factors are the factors for the conventional
units.

Users will want to work with the conventionalRatio, but we'll
always prefer to encode as message-rate. SimpleOrder is updated to
reflect that.

+++ client/core/types.go
@@ -494,11 +494,11 @@ func newDisplayIDFromSymbols(base, quote string) string {
 // SimpleOrder is minimal information about an order in a market's order book.
 // Replaced MiniOrder, which had a float Qty in conventional units.
 type SimpleOrder struct {
-	Qty   uint64  `json:"qty"`
-	Rate  float64 `json:"rate"`
-	Epoch uint64  `json:"epoch,omitempty"`
-	Sell  bool    `json:"sell"`
-	Token string  `json:"token"`
+	Qty   uint64 `json:"qty"`
+	Rate  uint64 `json:"rate"`
+	Epoch uint64 `json:"epoch,omitempty"`
+	Sell  bool   `json:"sell"`
+	Token string `json:"token"`
 }
 
 // RemainderUpdate is an update to the quantity for an order on the order book.
